Add files via upload
A chatbot interacts on a format similar to instant messaging. By artificially replicating the patterns of human interactions in machine learning allows computers to learn by themselves without programming natural language processing. A chatbot is programmed to work independently from a human operator. It can answer questions formulated to it in natural language and respond like a real person. It provides responses based on a combination of predefined scripts and machine learning applications.